fix(web): extract buildIdOnlyFrame out of stream route file
Same issue as sources/route.ts — Next.js rejects non-handler named exports
from route files. Move buildIdOnlyFrame + NotifyMessage + StreamChannel to
lib/security/stream-utils.ts; update route + test imports.
