+++ apps/web/src/app/api/monitoring/security/stream/route.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi } from 'vitest'
-import { buildIdOnlyFrame, type NotifyMessage } from './route'
+import { buildIdOnlyFrame, type NotifyMessage } from '@/lib/security/stream-utils'
 
 describe('stream/route', () => {
   describe('buildIdOnlyFrame', () => {
